chore: move `navbar.jsx` to section folder
