+++ app/page.jsx
@@ -2,7 +2,7 @@ import About from "@/components/section/about";
 import Footer from "@/components/section/footer";
 import Landing from "@/components/section/landing";
 import ProjectsGallery from "@/components/section/projects-gallery";
-import NavBar from "@/components/ui/navbar";
+import NavBar from "@/components/section/navbar";
 
 export default function Home() {
   return (
