+++ app/shara-clarke/page.jsx
@@ -1,7 +1,7 @@
 import ProjectInfo from "@/components/section/project-info";
 import React from "react";
 import projectData from "@/data/projects.json";
-import NavBar from "@/components/ui/navbar";
+import NavBar from "@/components/section/navbar";
 import Footer from "@/components/section/footer";
 import { ProjectNavigation } from "@/components/ui/project-navigation";
