+++ components/section/navbar.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import navbarData from "@/data/navbar";
-import { ZekeLogoIcon } from "./svg";
-import { Separator } from "./separator";
+import { ZekeLogoIcon } from "../ui/svg";
+import { Separator } from "../ui/separator";
 import { ModeToggle } from "@/components/ui/theme/mode-toggle";
 import { Sheet, SheetContent, SheetTrigger } from "@/components/ui/sheet";
 import Link from "next/link";
